Rename spark common integration tests
