+++ integration-test/spark-common/src/test/java/org/apache/uniffle/test/DynamicFetchClientConfIT.java
@@ -36,7 +36,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class DynamicFetchClientConfTest extends IntegrationTestBase {
+public class DynamicFetchClientConfIT extends IntegrationTestBase {
 
   @Test
   public void test() throws Exception {
