+++ integration-test/spark-common/src/test/java/org/apache/uniffle/test/SparkClientWithLocalIT.java
@@ -52,7 +52,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.fail;
 
-public class SparkClientWithLocalTest extends ShuffleReadWriteBase {
+public class SparkClientWithLocalIT extends ShuffleReadWriteBase {
 
   private static final String EXPECTED_EXCEPTION_MESSAGE = "Exception should be thrown";
   private static File DATA_DIR1;
